clang tidy

+++ Frontend/Draw/Draw2.hpp
@@ -1,14 +1,15 @@
 #include "Draw2.h"
 #include "App/GraphicsHandler.h"
+#include "Color/Color.h"
 
 inline void Draw2::Rectangle(const GraphicsHandler& handler,
                              const Rect<unsigned int>& rect,
                              const Color& color) {
    RECT reg{};
-   reg.left = rect.GetLeft();
-   reg.right = rect.GetRight();
-   reg.bottom = rect.GetBottom();
-   reg.top = rect.GetTop();
+   reg.left = static_cast<long>(rect.GetLeft());
+   reg.right = static_cast<long>(rect.GetRight());
+   reg.bottom = static_cast<long>(rect.GetBottom());
+   reg.top = static_cast<long>(rect.GetTop());
 
    FillRect(handler.GetBase(), &reg, CreateSolidBrush(color.ToPlatform()));
 }
@@ -18,7 +19,7 @@ void Draw2::Text(const GraphicsHandler& handler,
                  const Rect<unsigned int>& rect,
                  const Color& color,
                  const std::string& text) {
-   MoveToEx(handler.GetBase(), rect.GetX(), rect.GetY(), NULL);
+   MoveToEx(handler.GetBase(), static_cast<int>(rect.GetX()), static_cast<int>(rect.GetY()), NULL);
    SetTextAlign(handler.GetBase(), TA_CENTER); // TODO Parameter for text alignement
    SetTextColor(handler.GetBase(), color.ToPlatform());
    SetBkMode(handler.GetBase(), TRANSPARENT);
